+++ pilz_control/include/pilz_control/pilz_joint_trajectory_controller.h
@@ -176,6 +176,8 @@ protected:
 
   void triggerMovementToHoldPosition();
 
+  void trajectoryCommandCB(const JointTrajectoryConstPtr& msg) override;
+
 private:
   /**
    * @brief Invoke cartesian speed monitoring and perform controlled stop in case of speed limit violation.