+++ pilz_control/include/pilz_control/pilz_joint_trajectory_controller_impl.h
@@ -17,6 +17,8 @@
 #ifndef PILZ_CONTROL_PILZ_JOINT_TRAJECTORY_CONTROLLER_IMPL_H
 #define PILZ_CONTROL_PILZ_JOINT_TRAJECTORY_CONTROLLER_IMPL_H
 
+#include <string>
+
 #include <joint_trajectory_controller/joint_trajectory_segment.h>
 #include <joint_trajectory_controller/tolerances.h>
 
@@ -24,6 +26,19 @@ namespace pilz_joint_trajectory_controller
 {
 static constexpr double SPEED_LIMIT_ACTIVATED{ 0.25 };
 static constexpr double SPEED_LIMIT_NOT_ACTIVATED{ -1.0 };
+
+static const std::string USER_NOTIFICATION_NOT_IMPLEMENTED_COMMAND_INTERFACE_WARN{
+  "The topic interface of the original `joint_trajectory_controller` is deactivated. Please use the action interface "
+  "to send goals, that allows monitoring and receiving notifications about cancelled goals. If nonetheless you need "
+  "the topic interface feel encouraged to open an issue with this feature request at "
+  "https://github.com/PilzDE/pilz_robots/issues so that we can improve your user experience with our product."
+};
+
+static const std::string USER_NOTIFICATION_NOT_IMPLEMENTED_COMMAND_INTERFACE_INFO{
+  "For the reason behind the deactivation of this interface see "
+  "https://github.com/ros-controls/ros_controllers/issues/493). "
+  "PR welcome ;-)"
+};
 
 static const std::string LIMITS_NAMESPACE{ "limits" };
 
@@ -264,7 +279,7 @@ inline void PilzJointTrajectoryController<SegmentImpl, HardwareInterface>::updat
   {
     case TrajProcessingMode::unhold:
     {
-      if (!isPlannedUpdateOK(time_data.period) && mode_->stopEvent())
+      if (!isPlannedCartesianVelocityOK(time_data.period) && mode_->stopEvent())
       {
         stopMotion(time_data.uptime);
       }
@@ -367,6 +382,14 @@ bool PilzJointTrajectoryController<SegmentImpl, HardwareInterface>::handleMonito
   cartesian_speed_limit_ = req.data ? SPEED_LIMIT_ACTIVATED : SPEED_LIMIT_NOT_ACTIVATED;
   res.success = true;
   return true;
+}
+
+template <class SegmentImpl, class HardwareInterface>
+void PilzJointTrajectoryController<SegmentImpl, HardwareInterface>::trajectoryCommandCB(
+    const JointTrajectoryConstPtr& /*msg*/)
+{
+  ROS_WARN_STREAM_NAMED(this->name_, USER_NOTIFICATION_NOT_IMPLEMENTED_COMMAND_INTERFACE_WARN);
+  ROS_INFO_STREAM_NAMED(this->name_, USER_NOTIFICATION_NOT_IMPLEMENTED_COMMAND_INTERFACE_INFO);
 }
 
 }  // namespace pilz_joint_trajectory_controller